Refactored top-tracks out of main for easier testing and maintainence
